Refactor types into organized directory
Refactor type definitions into a `types` directory with organized subdirectories.

* Create a `types` directory with subdirectories for `user`, `post`, `tag`, `comment`, `notification`, `subscription`, `tipp`, `community`, and `files`.
* Move type definitions from `app/types.ts` to respective files in the `types` directory.
* Update import statements in `app/layout.tsx` to reflect the new type locations.
* Delete `app/types.ts`.
* Add missing type references to `app/schema.sql`.
