+++ types/community.ts
@@ -1,0 +1,9 @@
+export interface Community {
+    id: string;
+    name: string;
+    description?: string;
+    created_by: string;
+    created_at: string;
+    private: boolean;
+    tenant_id: string;
+}
