+++ types/files.ts
@@ -1,0 +1,8 @@
+export interface Files {
+    id: string;
+    user_id: string;
+    post_id: string;
+    bucket_name?: string;
+    file_path?: string;
+    tenant_id: string;
+}
